Moved csv pasing utils to parse_csv.py
